feat(Content-OC): Upgrade OC actions to a `TryGet` method, separate and organize the individual phantom jobs according to the CCP.
This removes OC's previous `ShouldUsePhantomActions` method, and upgrades the `BestPhantomAction` method to a `TryGet` method. Firstly to prevent all of the Phantom logic from running twice when an OC action is desired, and secondly to streamline the code needed within each job.
This also replaces the Phantom Jobs' `region`s with actual separate (`private`) methods for each Job, and organizes them all in the order in which they appear in the CCP file.